Even more drastic rearranging
